Fix empty string in multiple selector; Add helper text
